backend/requirements/development.txt
-r ./base.txt

django-extensions==3.1.5

pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.6.1

Django
mysqlclient
django-mysql
djangorestframework
ptvsd

numpy
pandas
django-cache-machine

psycopg2-binary
django-cors-headers
geojson
